+++ resources/views/components/tables/kpi-summary.blade.php
@@ -1,7 +1,0 @@
-@props(['cards'])
-
-<div class="tables-summary tables-summary--kpi" data-tables-summary="kpi">
-    @foreach ($cards as $card)
-        <x-tables.kpi-card :card="$card"/>
-    @endforeach
-</div>