+++ src/Summary/Summary.php
@@ -2,4 +2,10 @@
 
 namespace Mercurio\Tables\Summary;
 
-abstract class Summary {}
+final class Summary
+{
+    /**
+     * @param  array<int, SummaryCard>  $cards
+     */
+    public function __construct(public readonly array $cards) {}
+}