+++ src/Summary/SummaryCard.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Mercurio\Tables\Summary;
+
+abstract class SummaryCard
+{
+    abstract public function cellView(): string;
+}